arreglo teorema d epitagoras

+++ JavaFundamentals/teoremaPitagoras/Pitagoras.java
@@ -2,7 +2,7 @@ import java.lang.Math;
 
 public class Pitagoras {
 
-    public double calcularHipotenusa(int catetoA, int catetoB){
+    public double calcularHipotenusa(Double catetoA, Double catetoB){
         double c = (catetoA*catetoA) + (catetoB * catetoB);
         double result = Math.sqrt(c);
         return result;
